Класс TableFiles вынесен в отдельные файлы
