+++ src/ctool1cd/Class_1CD.h
@@ -13,6 +13,7 @@
 #include "Field.h"
 #include "Index.h"
 #include "Table.h"
+#include "TableFiles.h"
 
 //---------------------------------------------------------------------------
 
@@ -20,6 +21,8 @@ class T_1CD;
 class Table;
 class v8object;
 class Field;
+class table_file;
+class TableFiles;
 
 class ConfigStorageTableConfig;
 class ConfigStorageTableConfigSave;
@@ -83,62 +86,6 @@ union root
 {
 	root_80 root80;
 	root_81 root81;
-};
-
-//---------------------------------------------------------------------------
-// Структура адреса файла таблицы-контейнера файлов
-struct table_blob_file
-{
-	uint32_t blob_start;
-	uint32_t blob_length;
-};
-
-//---------------------------------------------------------------------------
-// Структура записи таблицы контейнера файлов
-struct table_rec
-{
-	String name;
-	table_blob_file addr;
-	int32_t partno;
-	FILETIME ft_create;
-	FILETIME ft_modify;
-};
-
-//---------------------------------------------------------------------------
-// Структура файла таблицы контейнера файлов
-struct table_file
-{
-	Table* t;
-	String name; // Имя, как оно хранится в таблице
-	uint32_t maxpartno;
-	table_blob_file* addr;
-	FILETIME ft_create;
-	FILETIME ft_modify;
-
-	table_file(Table* _t, const String& _name, uint32_t _maxpartno);
-	~table_file();
-};
-
-//---------------------------------------------------------------------------
-// Класс таблицы контейнера файлов (CONFIG, CONFIGSAVE, PARAMS, FILES, CONFICAS, CONFICASSAVE)
-class TableFiles
-{
-private:
-	Table* tab;
-	std::map<String,table_file*> allfiles;
-	char* rec;
-	bool ready;
-
-	bool test_table();
-public:
-	TableFiles(Table* t);
-	virtual ~TableFiles();
-	bool getready(){return ready;};
-	table_file* getfile(const String& name);
-	Table* gettable(){return tab;};
-
-	// __property std::map<String,table_file*> files = {read = allfiles};
-	std::map<String,table_file*> &files();
 };
 
 //---------------------------------------------------------------------------
